Changes to be committed: 	new file:   1_Rayleigh_and_Rice/PSD_test.csv 	new file:   1_Rayleigh_and_Rice/PSDtest.cpp 	modified:   1_Rayleigh_and_Rice/SpectralProperties.cpp 	modified:   1_Rayleigh_and_Rice/SpectralProperties.hpp

+++ 1_Rayleigh_and_Rice/SpectralProperties.hpp
@@ -11,11 +11,11 @@ public:
 
   //also inplement: the Oth-order Bessel function of the first kind.
 
-  static double jakesPSD(long long int freq, double std_dev, double fmax /*Maximum Doppler frequency*/);
+  static double jakesPSD(double freq, double std_dev, double fmax /*Maximum Doppler frequency*/);
   static double jakesACF(short int tau, double std_dev, double fmax /*Maximum Doppler frequency*/);
   static double jakesBeta(double std_dev, double fmax /*Maximum Doppler frequency*/);
 
-  static double GaussianPSD(long long int freq, double std_dev, double fc /*3-dB-cut-off frequency*/);
+  static double GaussianPSD(double freq, double std_dev, double fc /*3-dB-cut-off frequency*/);
   static double GaussianACF(short int tau, double std_dev, double fc /*3-dB-cut-off frequency*/);
   static double GaussianBeta(double std_dev, double fc /*3-dB-cut-off frequency*/);
 };
